+++ src/server/formatter.ts
@@ -1,9 +1,11 @@
 import { TextDocument, TextEdit } from "vscode-languageserver-textdocument";
 import {
+  BinaryExpressionNode,
   NullLiteralNode,
   SourceDocumentNode,
   TableLiteralNode,
 } from "../parser";
+import { VariableAssignmentKinds } from "../parser/token-kind";
 
 export type FormatterOptions = {
   "null-literal": "null" | "nil" | "any";
@@ -104,6 +106,56 @@ export class Formatter {
               start: document.positionAt(node.rightDelimiter.start),
               end: document.positionAt(
                 node.rightDelimiter.start + node.rightDelimiter.length,
+              ),
+            },
+          });
+        }
+        return;
+      }
+
+      if (element.kind === "BinaryExpressionNode") {
+        const node = element as BinaryExpressionNode;
+
+        if (!VariableAssignmentKinds.includes(node.operator.kind)) {
+          return;
+        }
+
+        let text: string | undefined;
+
+        if (
+          node.operator.kind === "EqualsOperator" &&
+          options["assignment-token"] === "colon"
+        ) {
+          text = ":";
+        } else if (
+          node.operator.kind === "ColonOperator" &&
+          options["assignment-token"] === "equals"
+        ) {
+          text = "=";
+        } else if (
+          options["assignment-token"] === "table-colon-function-equals"
+        ) {
+          const isTableAssignment =
+            node.operator.getParentOfKind("ExpressionStatementNode")?.parent
+              ?.kind === "TableLiteralNode";
+
+          if (isTableAssignment && node.operator.kind === "EqualsOperator") {
+            text = ":";
+          } else if (
+            !isTableAssignment &&
+            node.operator.kind === "ColonOperator"
+          ) {
+            text = "=";
+          }
+        }
+
+        if (text) {
+          edits.push({
+            newText: text,
+            range: {
+              start: document.positionAt(node.operator.start),
+              end: document.positionAt(
+                node.operator.start + node.operator.length,
               ),
             },
           });